requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
Pillow==10.0.0
PyPDF2==3.0.1
fpdf==1.7.2
jinja2==3.1.2
aiofiles==23.1.0
python-dotenv==1.0.0